# contracts/api/v2/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminUserCollection": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "description": "The page's items; empty, never null",
            "items": {
              "$ref": "#/components/schemas/AdminUser"
            },
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Cursor state; absent for bounded unpaginated collections"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "DeviceLoginDecision": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "description": "State after the decision",
            "enum": [
              "approved",
              "denied"
            ],
            "examples": [
              "approved"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "Safe occurrence-specific explanation; not for control flow",
            "examples": [
              "The request did not pass validation; see errors."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Field-level validation details",
            "items": {
              "$ref": "#/components/schemas/ProblemError"
            },
            "type": "array"
          },
          "instance": {
            "description": "urn:silo:request:\u003crequest-id\u003e, matching the X-Request-ID response header",
            "examples": [
              "urn:silo:request:000000000000000000000003"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, equal to the response status",
            "examples": [
              422
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "Short summary fixed for the problem type",
            "examples": [
              "Validation failed"
            ],
            "type": "string"
          },
          "type": {
            "description": "Stable problem type URI; the final segment is the problem identifier",
            "examples": [
              "https://siloserver.org/docs/api/v2/problems/validation_failed"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A session token or an API key in the Authorization header. The server decides which it holds.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Silo API",
    "version": "2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/admin/users": {
      "get": {
        "operationId": "listAdminUsers",
        "parameters": [
          {
            "description": "Optional. When present, it must name the authenticated account's primary profile; an absent header is accepted.",
            "in": "header",
            "name": "X-Profile-Id",
            "schema": {
              "examples": [
                "1"
              ],
              "type": "string"
            }
          },
          {
            "description": "Verification proof for a PIN-locked profile, issued by POST /api/v1/profiles/{id}/verify-pin until that operation moves to v2; required only when the declared profile is locked",
            "in": "header",
            "name": "X-Profile-Token",
            "schema": {
              "examples": [
                "pvt_5f3a9c1e7b2d4e8fa0c6"
              ],
              "type": "string"
            }
          },
          {
            "description": "Page size; default 50, maximum 200",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 50,
              "description": "Page size; default 50, maximum 200",
              "examples": [
                50
              ],
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from page.next_cursor",
            "explode": false,
            "in": "query",
            "name": "cursor",
            "schema": {
              "description": "Opaque cursor from page.next_cursor",
              "examples": [
                "eyJpIjo1MH0"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserCollection"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List login accounts with their policy overrides and effective policy, in account id order.",
        "tags": [
          "admin"
        ],
        "x-silo-class": "acting_admin",
        "x-silo-demo-restricted": true,
        "x-silo-service-backed": true
      }
    },
...
    "/api/v2/auth/device/approve": {
      "post": {
        "operationId": "approveDeviceLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideDeviceLoginInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceLoginDecision"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Gone"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Approve a pairing request as the caller's account.",
        "tags": [
          "device-login"
        ],
        "x-silo-class": "authenticated",
        "x-silo-service-backed": true
      }
    },
...
    "/api/v2/auth/device/poll": {
      "post": {
        "operationId": "pollDeviceLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PollDeviceLoginInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceLoginPoll"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "406": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Acceptable"
          },
          "408": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Timeout"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Poll a pairing request from the device and collect its tokens once approved.",
        "tags": [
          "device-login"
        ],
        "x-silo-class": "public",
        "x-silo-service-backed": true
      }
    },
...
  }
}